data_processor/requirements.txt
paho-mqtt
psycopg2


